requirements.txt
```txt
streamlit
google-generativeai
python-dotenv
speechrecognition
pydub